Added pKa and Hydro. module
